# src/main/resources/static/docs/openapi3.yaml
openapi: 3.0.1
info:
  title: API 문서
  description: RestDocs With Swagger
  version: 0.0.1
servers:
- url: http://devterview.com:8080
- url: http://localhost:8080
- url: http://dev-api.devterview.com:8080
- url: http://dev-api.devterview.com
...
paths:
...
  /post:
    post:
      tags:
      - Post API
      summary: 게시글 API
      description: 게시글 API
      operationId: 게시글 작성 테스트(
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/post-881832127'
            examples:
              게시글 작성 테스트(면접 공유 INTERVIEW) - 성공:
                value: |-
                  {
                    "jobKoreanName" : "백엔드",
                    "content" : "This is a test post content.",
                    "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                    "interviewId" : 1,
                    "overallEvaluationId" : 2,
                    "postType" : "INTERVIEW"
                  }
              게시글 작성 테스트(일반 POST) - 성공:
                value: |-
                  {
                    "jobKoreanName" : "백엔드",
                    "content" : "This is a test post content.",
                    "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                    "interviewId" : 1,
                    "overallEvaluationId" : 2,
                    "postType" : "POST"
                  }
              게시글 작성 테스트(평가 공유 EVALUATION) - 성공:
                value: |-
                  {
                    "jobKoreanName" : "백엔드",
                    "content" : "This is a test post content.",
                    "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                    "interviewId" : 1,
                    "overallEvaluationId" : 2,
                    "postType" : "EVALUATION"
                  }
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/post-218793856'
              examples:
                게시글 작성 테스트(면접 공유 INTERVIEW) - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "postId" : 100,
                        "authorId" : 12,
                        "authorNickname" : "nickname",
                        "authorProfileUrl" : "authorProfileUrl",
                        "jobName" : "BACK_END",
                        "jobNameKorean" : "백엔드",
                        "content" : "This is a test post content.",
                        "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                        "interview" : {
                          "interviewId" : 1,
                          "interviewTitle" : "Mock Interview Title",
                          "interviewStatus" : "COMPLETED",
                          "interviewType" : "TECHNICAL",
                          "interviewMethod" : "VIDEO",
                          "interviewMode" : "REAL",
                          "questionCount" : 10,
                          "job" : {
                            "jobId" : 1,
                            "jobName" : "BACK_END",
                            "jobNameKorean" : "백엔드",
                            "jobDescription" : "백엔드 직무입니다."
                          },
                          "files" : [ {
                            "fileId" : 1,
                            "type" : "COVER_LETTER",
                            "fileName" : "cover_letter.pdf",
                            "s3FileUrl" : "https://s3.amazonaws.com/resume.pdf"
                          } ]
                        },
                        "evaluation" : null,
                        "postType" : "INTERVIEW",
                        "generatedAt" : "2024-12-19T16:35:56.146705"
                      }
                    }
                게시글 작성 테스트(일반 POST) - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "postId" : 100,
                        "authorId" : 12,
                        "authorNickname" : "nickname",
                        "authorProfileUrl" : "authorProfileUrl",
                        "jobName" : "BACK_END",
                        "jobNameKorean" : "백엔드",
                        "content" : "This is a test post content.",
                        "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                        "interview" : null,
                        "evaluation" : null,
                        "postType" : "POST",
                        "generatedAt" : "2024-12-19T16:35:56.103629"
                      }
                    }
                게시글 작성 테스트(평가 공유 EVALUATION) - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "postId" : 100,
                        "authorId" : 12,
                        "authorNickname" : "nickname",
                        "authorProfileUrl" : "authorProfileUrl",
                        "jobName" : "BACK_END",
                        "jobNameKorean" : "백엔드",
                        "content" : "This is a test post content.",
                        "s3ImageUrl" : "https://s3.amazonaws.com/test-image.jpg",
                        "interview" : null,
                        "evaluation" : {
                          "interview" : {
                            "interviewId" : 1,
                            "interviewTitle" : "Mock Interview Title",
                            "interviewStatus" : "COMPLETED",
                            "interviewType" : "TECHNICAL",
                            "interviewMethod" : "VIDEO",
                            "interviewMode" : "REAL",
                            "questionCount" : 10,
                            "job" : {
                              "jobId" : 1,
                              "jobName" : "BACK_END",
                              "jobNameKorean" : "백엔드",
                              "jobDescription" : "백엔드 직무입니다."
                            },
                            "files" : [ {
                              "fileId" : 1,
                              "type" : "COVER_LETTER",
                              "fileName" : "cover_letter.pdf",
                              "s3FileUrl" : "https://s3.amazonaws.com/resume.pdf"
                            } ]
                          },
                          "evaluationCriteria" : [ {
                            "evaluationCriteriaId" : 1,
                            "evaluationCriteria" : "Communication Skills",
                            "feedbackText" : "Clear and concise.",
                            "score" : 9
                          }, {
                            "evaluationCriteriaId" : 2,
                            "evaluationCriteria" : "Problem-Solving",
                            "feedbackText" : "Good examples provided.",
                            "score" : 8
                          } ],
                          "answerEvaluations" : [ {
                            "answerEvaluationId" : 1,
                            "questionText" : "What is your greatest strength?",
                            "answerText" : "I am highly adaptable.",
                            "answerS3PresignedUrl" : "https://www.bucket.com",
                            "answerFeedbackStrength" : "Good emphasis on adaptability.",
                            "answerFeedbackImprovement" : "Could elaborate more.",
                            "answerFeedbackSuggestion" : "Provide an example of adaptability.",
                            "answerEvaluationScores" : [ {
                              "answerEvaluationScoreId" : 1,
                              "answerEvaluationScoreName" : "Clarity",
                              "score" : 8,
                              "rationale" : "Clear explanation",
                              "answerEvaluationType" : "TEXT"
                            }, {
                              "answerEvaluationScoreId" : 2,
                              "answerEvaluationScoreName" : "Conciseness",
                              "score" : 7,
                              "rationale" : "Could be more concise",
                              "answerEvaluationType" : "TEXT"
                            } ]
                          } ]
                        },
                        "postType" : "EVALUATION",
                        "generatedAt" : "2024-12-19T16:35:56.057233"
                      }
                    }
  /post/search:
    get:
      tags:
      - Post API
      summary: 게시글 검색 API
      description: 게시글 검색 API
      operationId: 게시글 검색 - 성공
      parameters:
      - name: keyword
        in: query
        description: 검색 키워드
        required: true
        schema:
          type: string
      - name: page
        in: query
        description: 페이지 번호 (0부터 시작)
        required: true
        schema:
          type: string
      - name: size
        in: query
        description: 페이지당 게시글 수
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/post-search1283205629'
              examples:
                게시글 검색 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "posts" : [ {
                          "postId" : 100,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is a test post about Spring Boot.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image1.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113721"
                        }, {
                          "postId" : 101,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Testing JPA and database connections.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image2.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113724"
                        }, {
                          "postId" : 102,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Unit test examples with JUnit and Mockito.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image3.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113727"
                        }, {
                          "postId" : 103,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Integration testing best practices.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image4.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113728"
                        }, {
                          "postId" : 104,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Test-driven development workflow.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image5.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113729"
                        }, {
                          "postId" : 105,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Testing RESTful APIs with Spring Boot Test.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image6.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113731"
                        }, {
                          "postId" : 106,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Performance testing using JMeter.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image7.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113732"
                        }, {
                          "postId" : 107,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "End-to-end testing strategies.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image8.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113733"
                        }, {
                          "postId" : 108,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Test coverage analysis with JaCoCo.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image9.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113735"
                        }, {
                          "postId" : 109,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "Automated testing pipeline setup.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image10.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.113736"
                        } ],
                        "currentPage" : 0,
                        "isLastPage" : false
                      }
                    }
  /post/subscription:
    get:
      tags:
      - Post API
      summary: 구독 게시글 API
      description: 구독 게시글 API
      operationId: 구독한 직무의 게시글 목록 페이징 조회 - 성공
      parameters:
      - name: page
        in: query
        description: 페이지 번호 (0부터 시작)
        required: true
        schema:
          type: string
      - name: size
        in: query
        description: 페이지당 게시글 수
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/post-search1283205629'
              examples:
                구독한 직무의 게시글 목록 페이징 조회 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "posts" : [ {
                          "postId" : 100,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 1.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image1.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081961"
                        }, {
                          "postId" : 101,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 2.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image2.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081965"
                        }, {
                          "postId" : 102,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 3.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image3.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081966"
                        }, {
                          "postId" : 103,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 4.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image4.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081968"
                        }, {
                          "postId" : 104,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 5.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image5.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.08197"
                        }, {
                          "postId" : 105,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 6.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image6.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081971"
                        }, {
                          "postId" : 106,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 7.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image7.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081973"
                        }, {
                          "postId" : 107,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 8.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image8.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081975"
                        }, {
                          "postId" : 108,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 9.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image9.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081977"
                        }, {
                          "postId" : 109,
                          "authorId" : 12,
                          "authorNickname" : "nickname",
                          "authorProfileUrl" : "authorProfileUrl",
                          "jobName" : "BACK_END",
                          "jobNameKorean" : "백엔드",
                          "content" : "This is test post content 10.",
                          "s3ImageUrl" : "https://s3.amazonaws.com/test-image10.jpg",
                          "interview" : null,
                          "evaluation" : null,
                          "postType" : "POST",
                          "generatedAt" : "2024-12-19T16:35:56.081978"
                        } ],
                        "currentPage" : 0,
                        "isLastPage" : true
                      }
                    }
...
  /subscription:
    post:
      tags:
      - Subscription API
      summary: 구독 생성 API
      description: 구독 생성 API
      operationId: 구독 생성 테스트 - 성공
      requestBody:
        content:
          application/json;charset=UTF-8:
            schema:
              $ref: '#/components/schemas/subscription-2024782262'
            examples:
              구독 생성 테스트 - 성공:
                value: |-
                  {
                    "jobId" : 2
                  }
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/subscription129725505'
              examples:
                구독 생성 테스트 - 성공:
                  value: |-
                    {
                      "code" : 200,
                      "message" : "SUCCESS",
                      "data" : {
                        "subscriptionId" : 1,
                        "userId" : 1,
                        "jobId" : 2,
                        "subscribedAt" : "2024-12-19T16:35:58.225611"
                      }
                    }
...
components:
  schemas:
    post-search1283205629:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            isLastPage:
              type: boolean
              description: 마지막 페이지 여부
            currentPage:
              type: number
              description: 현재 페이지 번호
            posts:
              type: array
              description: 게시글 목록
              items:
                type: object
                properties:
                  jobName:
                    type: string
                    description: 직무 이름
                  postType:
                    type: string
                    description: 게시글 유형
                  generatedAt:
                    type: string
                    description: 작성 일시
                  jobNameKorean:
                    type: string
                    description: 직무 이름 (한글)
                  s3ImageUrl:
                    type: string
                    description: S3 이미지 URL
                  postId:
                    type: number
                    description: 게시글 식별자
                  authorProfileUrl:
                    type: string
                    description: 작성자 프로필 이미지 URL
                  authorNickname:
                    type: string
                    description: 작성자 닉네임
                  authorId:
                    type: number
                    description: 작성자 식별자
                  content:
                    type: string
                    description: 게시글 내용
        message:
          type: string
          description: 응답 메시지
...
    subscription-2024782262:
      type: object
      properties:
        jobId:
          type: number
          description: 직무 ID
...
    post-881832127:
      type: object
      properties:
        overallEvaluationId:
          type: number
          description: 종합 평가 식별자
        interviewId:
          type: number
          description: 면접 식별자
        jobKoreanName:
          type: string
          description: 직무 이름 (한글)
        postType:
          type: string
          description: "게시글 유형 (EVALUATION, INTERVIEW, POST)"
        s3ImageUrl:
          type: string
          description: S3 이미지 URL
        content:
          type: string
          description: 게시글 내용
...
    subscription129725505:
      type: object
      properties:
        code:
          type: number
          description: 응답 코드
        data:
          type: object
          properties:
            jobId:
              type: number
              description: 직무 ID
            subscribedAt:
              type: string
              description: 구독 생성 시각
            subscriptionId:
              type: number
              description: 구독 ID
            userId:
              type: number
              description: 사용자 ID
        message:
          type: string
          description: 응답 메시지
  securitySchemes:
    APIKey:
      type: apiKey
      name: access_token
      in: cookie
security:
  - APIKey: []